I implemented the start of the Multiplayer game. I implemented websockets so that the server can send questions to the clients.
When a player joins the open lobby, he/she gets registered to the websocket channel of the current open lobby. When somebody presses play, the game starts and a new current open lobby is generated.

The server sends questions to the client. The client receives the questions and will react to them. For now, the client will only show the title of the first activity in the activity list, but this will be changed so that the client will show the right question format for each question. At this moment, the multiplayer game will not stop but will keep on going until the client closes the window. This will be changed so that the client will play only 20 questions